dentist_sms_system/requirements.txt
Flask==2.3.3
twilio==8.11.0
python-dotenv==1.0.0
black==23.9.1
schedule==1.2.0 